+++ PerebasFC/PerebasFC/Features/ParticipantsManager/Presenter/ParticipantsManagerPresenter.swift
@@ -67,7 +67,7 @@ final class ParticipantsManagerPresenter: ParticipantsManagerPresenterProtocol {
                         position: document["position"] as? String ?? "",
                         team: document["team"] as? String ?? "", 
                         warning: nil,
-                        rankingPosition: document["rankingPlace"] as? Int ?? 0,
+                        rankingPosition: document["rankingPlace"] as? Float ?? 0,
                         isAdm: Session.shared.isAdm ?? false,
                         type: ParticipantType(rawValue: document["type"] as? String ?? "") ?? .member,
                         menuItems: nil,
