+++ PerebasFC/PerebasFC/Features/Sort/Main/Worker/SortMainWorker.swift
@@ -183,7 +183,7 @@ final class SortMainWorker: SortMainWorkerProtocol {
                     position: document["position"] as? String ?? "",
                     team: document["team"] as? String ?? "",
                     warning: nil,
-                    rankingPosition: document["rankingPlace"] as? Int ?? 0,
+                    rankingPosition: document["rankingPlace"] as? Float ?? 0.0,
                     isAdm: Session.shared.isAdm ?? false,
                     type: ParticipantType(rawValue: document["type"] as? String ?? "") ?? .player,
                     menuItems: nil,
